requirements.txt
requests == 2.24.0
beautifulsoup4 == 4.9.3
google-api-python-client == 2.65.0
google-auth-httplib2 == 0.1.0
google-auth-oauthlib == 0.7.0
oauth2client == 4.1.3
google-cloud-bigquery == 3.3.5
cchardet == 2.1.7
lxml == 4.9.1
